+++ src/Framework/Node.cpp
@@ -188,7 +188,7 @@ void Node::ReloadProperties()
 void Node::ReloadPropertiesRecursive()
 {
 	ReloadProperties();
-	for (int i=0; i < childs.size(); i++) {
+	for (size_t i=0; i < childs.size(); i++) {
 		childs[i]->ReloadPropertiesRecursive();
 	}
 }
